// src/model/familyBased/biographic.ts
export interface Biographic {
  // Part3-1
  ethnicityHispanicCheckbox: string | undefined;
  ethnicityNotHispanicCheckbox: string | undefined;

  // Part3-2
  raceWhiteCheckbox: string | undefined;
  raceAsianCheckbox: string | undefined;
  raceBlackCheckbox: string | undefined;
  raceAmericanIndianCheckbox: string | undefined;
  raceNativeHawaiianCheckbox: string | undefined;

  // 268
  heightFeet: string;
  // 269
  heightInches: string;
  // 27
  weightPounds1: string;
  weightPounds2: string;
  weightPounds3: string;
  // Part3-5
  eyeColorBlackCheckbox: string | undefined;
  eyeColorBlueCheckbox: string | undefined;
  eyeColorBrownCheckbox: string | undefined;
  eyeColorGrayCheckbox: string | undefined;
  eyeColorGreenCheckbox: string | undefined;
  eyeColorHazelCheckbox: string | undefined;
  eyeColorMaroonCheckbox: string | undefined;
  eyeColorPinkCheckbox: string | undefined;
  eyeColorUnknownCheckbox: string | undefined;
  // Part3-6
  hairColorBaldCheckbox: string | undefined;
  hairColorBlackCheckbox: string | undefined;
  hairColorBlondCheckbox: string | undefined;
  hairColorBrownCheckbox: string | undefined;
  hairColorGrayCheckbox: string | undefined;
  hairColorRedCheckbox: string | undefined;
  hairColorSandyCheckbox: string | undefined;
  hairColorWhiteCheckbox: string | undefined;
  hairColorUnknownCheckbox: string | undefined;
}