// src/model/familyBased/entryInformation.ts
export interface EntryInformation {
  beneficiaryEverInUSYesCheckbox: string;
  beneficiaryEverInUSNoCheckbox: string;
  arrivedAsAdmission: string;
  i94Number: string;
  dateOfArrival: string;
  authorizedStayExpirationDate: string;
  passportNumber: string;
  travelDocumentNumber: string;
  passportIssueCountry: string;
  expirationDateForPassport: string;
}