// src/model/familyBased/intepreter.ts
export interface Interpreter {
  // 484
  lastName: string;
  // 483
  firstName: string;
  // 485
  businessName: string;
  // 487
  streetNumberAndName: string;
  // Part 7-3b 491
  aptCheckbox: string | undefined;
  steCheckbox: string | undefined;
  flrCheckbox: string | undefined;
  aptSteFlrNumber: string;
  // 486
  city: string;
  // 494
  state: string;
  // 493
  zipCode: string;
  // 496
  province: string;
  // 492
  postalCode: string;
  // 495
  country: string;
  // 503
  daytimeTelephoneNumber: string;
  // 505
  mobileTelephoneNumber: string;
  // 504
  emailAddress: string;
  // 497
  fluentInEnglish: string;
}