// src/model/familyBased/organization.ts
export interface Organization {
  orgName: string; // Used in P9Q2, P9Q6, P10Q10
  cityTown: string; // Used in P9Q3a, P9Q7a, P10Q11a
  state: string; // Used in P9Q3b, P9Q7b, P10Q11b
  country: string; // Used in P9Q3c, P9Q7c, P10Q11c
  natureOfGroup: string; // Used in P9Q4, P9Q8, P10Q12
  dateFrom: string; // Used in P9Q5a, P10Q9a, P10Q13a
  dateTo: string; // Used in P9Q5b, P10Q9b, P10Q13b
}