// src/model/familyBased/pertitionerOtherName.ts
export interface PetitionerOtherName {
  pageNumber: string;
  partNumber: string;
  itemNumber: string;

  // Family name for petitioner's other name used
  lastName: string;
  // Given name for petitioner's other name used
  firstName: string;
  // Middle name for petitioner's other name used
  middleName: string;
}